Add Mandi data fetching and saving features
This update includes new code for data fetching from the 'mandi' (a marketplace in India) and saving it to a database. It also introduces a new 'Mandi' table in the database schema. An automatic script has been added which can fetch and store 'mandi' data for set intervals. Minor text updates have also been incorporated.